+++ templates/__src/sass/base/_typography.scss
@@ -2,7 +2,7 @@ h1, h2, h3, h4, h5, h6{
   font-family: $font-family-headings;
 }
 body{
-  font-family: "MaisonNeue Book";
+  font-family: "Roboto Regular";
 }
 // body{
 //   font-size: $font-size-base;
