+++ templates/package.json
@@ -5,30 +5,35 @@
   "main": "main.js",
   "dependencies": {},
   "devDependencies": {
-    "@welance/welance-bs": "^0.6.1",
-    "babel-core": "^6.24.1",
-    "babel-loader": "^7.0.0",
+    "@welance/welance-bs": "^0.8.0",
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.2",
     "babel-polyfill": "^6.23.0",
-    "babel-preset-env": "^1.5.1",
-    "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "^0.28.3",
-    "extract-text-webpack-plugin": "^2.1.0",
+    "babel-preset-env": "^1.6.0",
+    "clean-webpack-plugin": "^0.1.16",
+    "copy-webpack-plugin": "^4.1.0",
+    "css-loader": "^0.28.7",
+    "extract-text-webpack-plugin": "^3.0.0",
     "favicons-webpack-plugin": "0.0.7",
-    "html-webpack-plugin": "^2.28.0",
-    "imagemin-webpack-plugin": "^1.4.4",
+    "html-webpack-plugin": "^2.30.1",
+    "imagemin-webpack-plugin": "^1.5.2",
     "jquery": "^3.2.1",
-    "object-fit-images": "^3.2.3",
     "node-sass": "^4.5.3",
-    "optimize-css-assets-webpack-plugin": "^1.3.2",
-    "postcss-loader": "^2.0.5",
-    "sass-loader": "^6.0.5",
+    "object-fit-images": "^3.2.3",
+    "postcss-loader": "^2.0.6",
+    "postcss-scss": "^1.0.2",
+    "sass-loader": "^6.0.6",
+    "sass-mq": "^3.3.2",
     "style-loader": "^0.18.1",
-    "webpack": "^2.6.1"
+    "webpack": "^3.6.0",
+    "webpack-bundle-analyzer": "^2.9.0",
+    "webpack-dev-server": "^2.9.1",
+    "webpack-merge": "^4.1.0"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "NODE_ENV='dev' env UV_THREADPOOL_SIZE=100 webpack  --display-error-details -d --watch",
-    "prod": "NODE_ENV='prod' env UV_THREADPOOL_SIZE=100  webpack -p --optimize-minimize;"
+    "dev": "NODE_ENV='dev' env UV_THREADPOOL_SIZE=100 webpack --display-error-details -d --watch  --config webpack/webpack.dev.js",
+    "prod": "NODE_ENV='prod' env UV_THREADPOOL_SIZE=100 webpack -p --config webpack/webpack.prod.js"
   },
   "repository": {
     "type": "git",
