+++ templates/webpack/webpack.dev.js
@@ -1,0 +1,19 @@
+const merge = require('webpack-merge');
+const common = require('./webpack.common.js');
+const ExtractTextPlugin = require('extract-text-webpack-plugin');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+
+module.exports = merge(common, {
+  devtool: 'inline-source-map',
+  devServer: {
+    contentBase: './dist'
+  },
+  plugins: [
+		//about SASS compilation
+		new ExtractTextPlugin({
+			filename: "styles/main.css"
+		})
+  ]
+});
+
+
